Networkign updates

+++ game/src/Framework/Network/Schemas/Vector3.ts
@@ -4,13 +4,13 @@ import {
 } from '@colyseus/schema';
 
 export class Vector3 extends Schema {
-  @type("int32")
+  @type("float32")
   x: number = 0;
 
-  @type("int32")
+  @type("float32")
   y: number = 0;
 
-  @type("int32")
+  @type("float32")
   z: number = 0;
 
   set(object: any) {
